Fixed Ball Sawnner
Balls are spawning from the controller. Fixed the issue with the target . Targets are spawning after being destroyed and after 10 seconds as ell if the ball do not hit the target. Score card also stopped working so fixed that as well. And also, fixed the issue with game over UI and Final Score UI.